Added Cancellators, IO-pipes that get written to upon termination
IO pipes are thread-safe inter-thread communication pipes. When an
Acceptor is created, a Cancellator is created and it's reading part
is stored in the Acceptor, its corresponding writer is stored in
the Stage. When the Stage is terminated, it writes to all the
Cancellators.
This is used to stop the blocking subscription in the
AMQPAcceptor to be able to perform a clean shutdown